[front] refactor(hubspot): migrate to new MCP server architecture
