+++ front/lib/actions/mcp_internal_actions/servers/index.ts
@@ -21,7 +21,6 @@ import { default as gmailServer } from "@app/lib/actions/mcp_internal_actions/se
 import { default as driveServer } from "@app/lib/actions/mcp_internal_actions/servers/google_drive";
 import { default as sheetsServer } from "@app/lib/actions/mcp_internal_actions/servers/google_sheets";
 import { default as httpClientServer } from "@app/lib/actions/mcp_internal_actions/servers/http_client";
-import { default as hubspotServer } from "@app/lib/actions/mcp_internal_actions/servers/hubspot";
 import { default as interactiveContentServer } from "@app/lib/actions/mcp_internal_actions/servers/interactive_content";
 import { default as jiraServer } from "@app/lib/actions/mcp_internal_actions/servers/jira";
 import { default as jitTestingServer } from "@app/lib/actions/mcp_internal_actions/servers/jit_testing";
@@ -60,6 +59,7 @@ import { default as extractDataServer } from "@app/lib/api/actions/servers/extra
 import { default as fileGenerationServer } from "@app/lib/api/actions/servers/file_generation";
 import { default as githubServer } from "@app/lib/api/actions/servers/github";
 import { default as calendarServer } from "@app/lib/api/actions/servers/google_calendar";
+import { default as hubspotServer } from "@app/lib/api/actions/servers/hubspot";
 import { default as imageGenerationServer } from "@app/lib/api/actions/servers/image_generation";
 import { default as includeDataServer } from "@app/lib/api/actions/servers/include_data";
 import { default as missingActionCatcherServer } from "@app/lib/api/actions/servers/missing_action_catcher";
@@ -116,7 +116,7 @@ export async function getInternalMCPServer(
     case "ashby":
       return ashbyServer(auth, agentLoopContext);
     case "hubspot":
-      return hubspotServer();
+      return hubspotServer(auth, agentLoopContext);
     case "image_generation":
       return imageGenerationServer(auth, agentLoopContext);
     case "speech_generator":
